Overhaul `mk1` to use a more efficient input format, relying upon an external like `tmpls` to do more pre-computation.  (This idea can be extended from 2-tuples to 4-tuples with the 2 new slots being the quoted variant.  A stat-input/print-output bit picks, essentially.)
Add support for sys_batch (with a `$BAT` env.var pointing to the repo
and a need to `nim c -d:batch`, and, of course, inserting the kernel
module to use the non-emulated wrapper).  Over 2X speed-up on SkyLake.
Include a hidden `--batch` option, but much more than 20-30 does not
seem helpful.

In combination, the first 2 edits induce a string-management logic
overhaul where strings are first saved in a re-locatable buffer with
`MSlice` pointers into that using "0-origin" pointer offsets.  Before
we want them to point to the real .mem, we just call `relTo` to make
them relative to the real base-address of `buf`.

Make `mk1 -B` 2X faster with faster single quoting logic and less
wasteful cInterPrint via new cligen/osUt.urite.

Add " #forced" explanation for `--always-make` & `--what-if` mode.

Most of the other logic stays the same, but the need to batch induces
splitting it into a pre-stat loop and body-template (to replicate the
same body post-loop).

--

Also update usage doc to match program.  Re-run bench to get similar,
but distinct results & update doc accordingly.  Trim the output a bit
with `ru -ht` and mention `sys_batch` in footnote.