Replace Einmahl 2009/2010 endpoint estimator with the Fraga Alves & Neves 2014/2017 Estimator for Right Endpoint and hokey re-sampling idea with a more proper bootstrap standard error estimation.
Change CLI quite a bit in the process, but the old way was unstable in
too many ways anyway.  This one might actually be useful to someone who
isn't me.

Also update `doc/eve.md`, leaving the old Einmahl reference in it.

The bottom of this new bu/eve.nim mentions a pretty new paper that may
really be the ultimate estimator of ultimates, but this IID variant is
nice enough to push right now.