Do statistically principled ranges using `bu/eve.eLE` & `bu/eve.eRE` only when `k != 0`.  Eve-based data range estimates make `iterator edf` yield first & last data points at the edge estimates.  This makes plots show data range estimates by the left-right extent of color.  Multiple data sets will likely overlap at P=0,1 => overlay order/transparency do still matter.  This update makes these graphic nearly fully informative { missing only data range uncertainty estimates on one/both sides which could perhaps be handled with "(.." and "..)" }.
Handle semi-infinite range estimates 2 diff ways: either A) return IEEE
+-inf OR B) extend range by the same orderStat[k]-based x-val distance
used to estimate the same extreme edge.  Decide A|B based on the SIGN
of the alpha level for infinite tail testing given { using abs(that
value) in the actual test }.  Right now default to A as +-inf does not
seem to make gnuplot choke, but flipping this default to B for infinite
tails may be safer/better.  Stay tuned.

As part of this work also make very simple changes:
  - Rename first param of blur/tube from k for Kind to b for BandKind
  - Update blur & tube to take new `k` & `tailA` passed from CLI
  - Add `tailA` CLI pass through parameter; Smaller => fewer inf tails
  - Rename `tp` -> `fp` to free up the most natural `--tailA`
  - Shrink 5% extra y-space to only 3% which seems enough
  - Add documentation for new CLI options