This is a provisional testing commit to do benchmarking on an external to all this `T=string` version of the same program.
Essentially the problem was that, unless a proc is annotated with
`{.nodestroy.}`, Nim emits a destructor to clean-up prior to a set of
an already allocated seq slot, but in this case said slot could have
weakly deterministic garbage being `setLenUninit`.  stdlib also uses
`{.nodestroy.}` to work around this.

The "`=dup` it" call here is needed to avoid every `seq` slot sharing
the same `string` object which seems like some kind of move/refcount
optimization (which cannot be worked around with a `move` call since
the procedure parameter was not `var`).  This bit feels more like an
"unfortunate &| confusing property" than a Nim bug, per se.