`setLenUninit` seems really not an important optimization for this program (see below) and so this commit undoes the last one, but is left in version control history for future reference.
This program (new current HEAD) always worked with `setLenUninit` on
memory-mapped files (where no string duplication is needed, only the
value-type `MSlice` pointers into the map).

So, to assess the importance of `setLenUninit`, I compared `setLen` vs.
`setLenUninit` variants on a pessimal case of subsetting 99999 random
lines out of 100001 (from `seq 1 100001`).  This should maximize the
relative cost of extra zeroing in `setLen`.  Even so, it was barely
detectable.  Careful CPU-pinned `tim` runs showed only a (1.10 +- 0.51)%
slowdown from the initializing variant of `setLen` -- not statistically
significant.

The same rough scale of delta did show up across several different Ns
(9999, etc.) and so is probably still "real" (unsurprising, since it
does take non-zero time to zero memory).  Even so, a good deal of
finesse is needed to even measure this.  So, summarize it as "barely
even detectable in the worst case".  This worst case should also be a
very uncommon case for the reservoir sampling library, if it is ever
even used outside the CLI utility which is 50% IO/line parsing bound
even with a RAM filesystem.