In preparation for supporting `tail -n+2`, add custom type `NRow` and its arg parser of first-char-qualified integers which produces an enum tag & a number.
First use of this is to add new terminal-friendly mode where the number
is the number of head/tail sections which can fit on one terminal screen
without scrolling.  E.g., tails -t/3 a b would be about 1/3 of the
terminal height for each of the two files.  Since a common case might be
to do a fraction which will work for the argument list, a lone '/' is
interpreted as /numberOfArgs.  This is careful to count `eor` inside of
`header` & `sep` (which `~/.config/tails` may well set with """ to be
multi-line).  Along w/a colorizing `header= """ """` in the config, this
makes tails` a nicer terminal application than system `head` | `tail`.

Note that the `\n==> PATH <==\n` separator in OG head/tail was surely
always intended to be a very terminal visible demarcation, not useful
for post processing, but inverse or colored text is much more visible,
not really even needing that extra blank line which along with this
auto-sizing protocol can be repurposed to show bigger file snippets.

As part of this, use `$eor` instead of literal `'\n'` in header and
rename `doPrint` to more specific `doHeaders`.

Also, update `doc/tails.md` usage, but more elaboration should really
be added since probably few read these commit messages.

Still need to use random access on seekable inputs to not read whole
files, handle negative head/tail, bytes units, and a follow protocol
before system head/tail can be full replaced with this.