Pure `head` mode was already fast/low IO.  This patch achieves the same for pure tail mode.  Not sure how portable `memrchr` is in practice.
Perf in that case is now comparable to GNU tail for a 16 MB file:
  # seq 1 500000|sed 's/^/abcde fghi jkl mno pq r s/' > /dev/shm/j
  # tim "/bin/tail /dev/shm/j>/dev/null" "tail /dev/shm/j>/dev/null"
  104.14 +- 0.77 μs       (AlreadySubtracted)Overhead
  180.3 +- 1.8 μs /bin/tail /dev/shm/j>/dev/null
  244.3 +- 2.9 μs tail /dev/shm/j>/dev/null
(Seems for EOF small IO, stdio is 1.3x faster than mmap though in
other circumstances mmap is 1.3x faster than stdio..). { Point of this
is not hyper-optimization -- this is (mostly) a terminal UI & humans
can only read so fast -- but rather to not be egregiously inefficient
in simple cases so as to allow hard/symlinking system head/tail. }