Fix off-by-1 bug @ZoomRmc aka @indiscipline complained about (I think, terminals do vary), at the cost of an extra blank row sometimes.
At least this test no longer does any weird scroll droppings on `st`:
```Zsh
(for a in {a..c};{for n in {1..3};{for b in {A..C};echo $a$n$b} })|vip -n15
```
with both `-n15` and `-n30`.  (That's actually a pretty good test case
for almost all features, order, case-insensitivity <|> terminal height,
requested list size, page up/dn and pattern matching.  Maybe it should
go in the doc.md?)