Add --keep hook/plug-in/loadable module system for lazy/delayed tests.
Basically just uses `cligen/osUt` loader wrapper to get an `okx` tester
function.  The main work is adding & using `xmatch` here to postprocess
the `filterQuit` filtered list.

We update the denominator in "5/75" to reflect deleted items which
induces a need for a new global `dig`.

Also add an example `cdable` function to `doc/vip.md` and update the
docs to reflect this change.

Undone, but with a TODO left behind is the trickier case of 1-row at
a time or page-down scrolling.