Revisit ideas of using `okx` to per `putN` not at `filterQuit` time.
This mostly works except that cursoring over missing entries on the
first page is weird (just from non-highlighting of the selection,
actual selection seems ok) and Scroll/PageDown/Up do not work yet
(an XXX is left here for that; have a little patience.  I hope to
get to that soon..the  `- yO/h`-bounding stuff needs some thought).